Incorporated the example repository into the luawrapper repo.
Instead of just being a pair of headers and a README, the repo will now
incorporate the examples as tests, and have tests for each supported
version of Lua.

Since the repo is more complicated now, a .gitignore has been added, as
well as a .clang-format.

There are also some examples and tests that are not hooked up yet.